pyproject.toml
[project]
name = "lasr-stealth-evals"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Nikita Ostrovsky", email = "nik.ostro@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "autogen-agentchat>=0.4.9.2",
    "autogen-core>=0.4.9.2",
    "autogen-ext>=0.4.9.2",
    "datasets>=3.3.2",
    "inspect-ai>=0.3.66",
    "langchain>=0.3.20",
    "langchain-community>=0.3.19",
    "langchain-openai>=0.3.7",
    "openai>=1.64.0",
    "pre-commit>=4.1.0",
    "tqdm>=4.67.1",
    "wandb>=0.19.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.10.0",
    "plotly>=6.0.0",
    "pre-commit>=4.1.0",
    "pyright>=1.1.394",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.1",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
]

[tool.ruff]
line-length = 120

[tool.pyright]
reportMissingParameterType = true
reportUnknownParameterType = true
reportUnusedComment = false
reportPrivateImportUsage = false
